app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Default template for PDM package"
authors = [ {name = "dudugazuforget", email = "dudu99744@gmail.com"},]
dependencies = [
    'django>=5.0.6',
    'djangorestframework>=3.14.0',
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.7.0",
    "shellp>=0.2.1",
    "shell>=1.0.1",
    "drf-spectacular>=0.28.0",
    "djangorestframework-simplejwt>=5.5.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
format = ["isort>=5.11.3", "black>=22.12.0"]
dev = [
    "black>=23.10.1",
    "isort>=5.12.0",
    "pylint>=3.0.2",
]

[tool.pdm.scripts]
dev = "python manage.py runserver"
createsuperuser = "python manage.py createsuperuser"
pre_migrate = "python manage.py makemigrations"
migrate = "python manage.py migrate"
shell = "python manage.py shell"
shellp= "python manage.py shell_plus"
test = "python manage.py test"
loaddata = "python manage.py loaddata core.json"
dumpdata = "python manage.py dumpdata --indent 2"
cria_api = "python ./scripts/cria_api.py {args}"
